# laravel-backend/public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "NessaPOS API",
    "description": "OpenAPI definition for the NessaPOS Laravel backend.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://api.domain.com/api"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/products/{product}": {
      "put": {
        "summary": "Update a product",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "product", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": {
          "200": { "description": "Product updated" }
        }
      },
      "delete": {
        "summary": "Delete a product",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "product", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": {
          "204": { "description": "Product deleted" }
        }
      }
    },
...
  }
}